workbench: compare a stored value against a binary32 domain (finding 52)
validateStoredValue rounded the value to binary32 but compared it against the
authored domain bounds. A catalog bound that is not float32-exact -- 64 of the
engine catalog's numeric bounds are not -- rejects a document that pins a value
to it, after the sibling domain-mismatch pass has already accepted the bound as
the catalog's own.

+++ scripts/shader_workbench.mjs
@@ -540,7 +540,11 @@ const validateStoredValue = (parameter, value, path) => {
   const stored = Math.fround(value);
   if (!Number.isFinite(stored))
     fail('semantic', 'BINARY32_OVERFLOW', path, 'The value is outside binary32 storage.');
-  if (stored < parameter.domain.minimum || stored > parameter.domain.maximum)
+  // The bounds are compared in binary32 too: a catalog bound the document
+  // copied verbatim need not be float32-exact, and the rounded value would
+  // then fall outside the unrounded domain.
+  if (stored < Math.fround(parameter.domain.minimum) ||
+      stored > Math.fround(parameter.domain.maximum))
     fail('semantic', 'VALUE_OUT_OF_RANGE', path, `The value is outside parameter "${parameter.id}" bounds.`);
   if (parameter.interpolation.kind === 'LOG_POSITIVE' && stored <= 0)
     fail('semantic', 'LOG_DOMAIN', path, 'Log-positive interpolation requires positive values.');
